first task

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -2,10 +2,10 @@
 #include "lists.h"
 
 /**
- * print_list - Print list function
+ * print_list - prints all the elements of a linked list
  * @h: pointer to the list_t list to print
  *
- * Return: returns the number of nodes printed
+ * Return: the number of nodes printed
  */
 size_t print_list(const list_t *h)
 {
@@ -16,7 +16,7 @@ size_t print_list(const list_t *h)
 		if (!h->str)
 			printf("[0] (nil)\n");
 		else
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%lu] %s\n", (unsigned long)h->len, h->str);
 		h = h->next;
 		s++;
 	}
